backend/requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
Django==4.1.2
django-cors-headers==3.13.0
djangorestframework==3.14.0
python-dotenv==0.21.0
pytz==2022.6
sqlparse==0.4.3
yapf==0.32.0
